servicio-be/requirements.txt
functions-framework==3.9.2
google-cloud-firestore==2.21.0
